首页iframe、俱乐部管理

+++ harem-biz/src/main/java/com/yimayhd/harem/service/impl/OrderServiceImpl.java
@@ -4,6 +4,7 @@ import com.yimayhd.harem.model.Commodity;
 import com.yimayhd.harem.model.Order;
 import com.yimayhd.harem.service.OrderService;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -18,15 +19,15 @@ public class OrderServiceImpl implements OrderService {
         order.setOrderNO("2015102710001");
         List<Commodity> commodityList = new ArrayList<Commodity>();
         Commodity commodity = new Commodity();
-        commodity.setOrderId(10086);
+        commodity.setOrderId(Long.valueOf(10086));
         commodity.setName("圆明园门票");
-        commodity.setPrice(15);
-        commodity.setNumber(2);
+        commodity.setPrice(BigDecimal.valueOf(15));
+        commodity.setNumber(Long.valueOf(2));
         Commodity commodity2 = new Commodity();
-        commodity2.setOrderId(10086);
+        commodity2.setOrderId(Long.valueOf(10086));
         commodity2.setName("故宫门票");
-        commodity2.setPrice(60);
-        commodity2.setNumber(2);
+        commodity2.setPrice(BigDecimal.valueOf(60));
+        commodity2.setNumber(Long.valueOf(2));
         commodityList.add(commodity);
         commodityList.add(commodity2);
         order.setCommodityList(commodityList);
